Better default ranging strategies
Uses RangingStrategy.default with 1 value per iteration for anytime algorithms, and ParticleGenerator.defaultNumSamplesFromAtomics values for one-time algorithms. Lazy algorithms use the same number of values, but instead use RangingStrategy.defaultLazy.
